finance-app/requirements.txt
streamlit==1.41.0
firebase-admin==6.6.0
pdfplumber==0.11.5
pandas==2.2.3
google-generativeai==0.8.3
python-dotenv==1.0.1
openpyxl==3.1.5

xlrd>=2.0.1
